backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation and utilities
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator>=2.1.0      # Required for pydantic EmailStr
python-dotenv==1.0.0
python-multipart==0.0.6

# AI Services
google-generativeai>=0.8.0
google-genai>=0.3.0
websockets>=12.0

# Audio Processing
numpy>=1.26.0

# Password hashing (direct, not through passlib)
bcrypt>=4.1.0

# Development
httpx==0.26.0
